+++ sample/core/src/lrs_waypoint_v2_sample.cpp
@@ -787,7 +787,7 @@ std::vector<WaypointV2>
 WaypointV2MissionSample::generateAngleWaypoints(float32_t step,
                                                 float32_t angle_deg,
                                                 uint16_t  n_points,
-                                                float32_t damping,
+                                                uint16_t damping,
                                                 float32_t speed)
 {
   // Let's create a vector to store our waypoints in.
